requirements.txt
alembic==1.12.1
fastapi==0.104.1
uvicorn==0.24.0.post1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0